moved compression to middlewares folder
