+++ bareclient/middlewares/__init__.py
@@ -1,0 +1,7 @@
+"""middlewares"""
+
+from .compression import compression_middleware
+
+__all__ = [
+    'compression_middleware'
+]
