+++ bareclient/middlewares/compression.py
@@ -1,22 +1,41 @@
 """Simple example with compression"""
 
-import asyncio
 from typing import (
     AsyncIterable,
     List,
+    Mapping,
     Optional,
     Tuple
 )
-from bareclient import (
-    HttpClient,
-    HttpClientCallback,
-    HttpClientMiddlewareCallback,
-    Response,
-    DEFAULT_DECOMPRESSORS,
-    DEFAULT_COMPRESSORS
+
+from bareutils import header
+from bareutils.compression import (
+    compression_reader_adapter,
+    compression_writer_adapter,
+    make_gzip_decompressobj,
+    make_deflate_decompressobj,
+    DecompressorFactory,
+    make_gzip_compressobj,
+    make_deflate_compressobj,
+    CompressorFactory
 )
-from bareutils import header, text_writer
-from bareutils.compression import compression_reader_adapter, compression_writer_adapter
+
+from ..types import Response
+from ..middleware import HttpClientCallback
+
+Decompressors = Mapping[bytes, DecompressorFactory]
+
+DEFAULT_DECOMPRESSORS: Decompressors = {
+    b'gzip': make_gzip_decompressobj,
+    b'deflate': make_deflate_decompressobj
+}
+
+Compressors = Mapping[bytes, CompressorFactory]
+
+DEFAULT_COMPRESSORS: Compressors = {
+    b'gzip': make_gzip_compressobj,
+    b'deflate': make_deflate_compressobj
+}
 
 
 def _make_body_writer(
@@ -58,29 +77,3 @@ async def compression_middleware(
     response = await handler(host, scheme, path, method, headers, content)
     response.body = _make_body_reader(response.headers, response.body)
     return response
-
-
-async def main(url: str) -> None:
-
-    headers = [
-        (b'content-type', b'text'),
-        (b'content-encoding', b'gzip'),
-        (b'accept-encoding', b'gzip')
-    ]
-    middleware: List[HttpClientMiddlewareCallback] = [
-        compression_middleware
-    ]
-    async with HttpClient(
-            url,
-            headers=headers,
-            middleware=middleware,
-            content=text_writer('Hello, World!')
-    ) as response:
-        print(response)
-        if response.status_code == 200 and response.body is not None:
-            async for part in response.body:
-                print(part)
-    print('Done')
-
-# asyncio.run(main('https://docs.python.org/3/library/cgi.html'))
-asyncio.run(main('https://beast.jetblack.net:9005/consume'))
