+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "bareclient"
-version = "4.6.0"
+version = "5.0.0-alpha.0"
 description = "A lightweight asyncio HTTP client"
 authors = ["Rob Blackbourn <rob.blackbourn@gmail.com>"]
 license = "Apache-2.0"
